+++ website/src/pages/elements/mod.rs
@@ -1,5 +1,7 @@
 mod button_page;
+mod icon_page;
 mod tag_page;
 
 pub use button_page::ButtonPage;
+pub use icon_page::IconPage;
 pub use tag_page::TagPage;